training/requirements.txt
tensorflow
tensorflowjs
numpy
Pillow
